1. 매입/매출리스트 조회 시 대수문자 구분 없이 조회 2. 제품그룹별 여신조회 신규컬럼 추가 3. 출하요청서 현장 등록시 바로 적용될 수 있도록 수정 4. 세금계산서 오류컬럼 삭제 기능 5. 에러세금계산서 삭제 기능 추가(유효성 검증)

+++ src/main/java/com/dksys/biz/user/ar/ar04/service/impl/AR04SvcImpl.java
@@ -787,9 +787,22 @@ public class AR04SvcImpl implements AR04Svc {
 	
 	@SuppressWarnings("all")
 	@Override
-	public void deleteTaxHd(Map<String, Object> paramMap) {
+	public void deleteTaxHd(Map<String, Object> paramMap) throws Exception {
+
+		// 세금계산서 에러 확인
+		Map<String, String> bilgParam = new HashMap<String, String>();
+		bilgParam.put("bilgCertNo", (String) paramMap.get("bilgCertNo"));
+		
+		Map<String, String> tax = ar04Mapper.selectTaxBilg(bilgParam);
+		
+		if(tax.get("cseqRcvYn") != "E") {
+			throw new Exception();
+		}
+		
+		// 세금계산서 삭제
 		ar04Mapper.updateAR02BilgCertNo(paramMap);
-		ar04Mapper.deleteTaxHd(paramMap);
+		ar04Mapper.deleteTaxHd(paramMap);	
+
 	}
 	
 }
